requirements.txt
pandas
numpy
scikit-learn
xgboost
prophet
cmdstanpy
streamlit
plotly
feedparser
requests
vaderSentiment
pyyaml
ipykernel
jupyter
black
ruff
pytest